Made it so that humdrum accounts can now be linked to spotify accounts somewhat directly
Spotify account names are stored in the database as SpotifyId in the user_pass table. The new things I added do an sql call for the current user then does a spotify search with the spotify usernae associated with the current user. It'll give errors for anyone without a spotify account linked